+++ src/Components/Routes.js
@@ -16,7 +16,7 @@ export default class Routes extends Component {
           <SiteRoute exact path="/sign-in" component={Site.SignIn} />
           <SiteRoute exact path="/sign-up" component={Site.SignUp} />
           <SiteRoute exact path="/pricing" component={Site.Pricing} />
-          <SiteRoute exact path="/create-question" component={Site.CreateQuestion} />
+          <SiteRoute exact path="/create-questionnaire" component={Site.CreateQuestionnaire} />
           <SiteRoute exact path="/my-questionnaires" component={Site.MyQuestionnaires} />
           <SiteRoute exact path="/about" component={Site.About} />
           <SiteRoute path="/polls/:pollId" component={Site.Polls} />
